Changed urealtime finish_time and start_timing to be int64_t, since they are now being saved in the save file. The same with time_stamp in save_game_stats. Changed also two size_t's in saved structs to uint64_t.

+++ include/context.h
@@ -109,7 +109,7 @@ struct obj_split {
 };
 
 struct tribute_info {
-    size_t tributesz;       /* make it possible to skip this in future */
+    uint64_t tributesz;       /* make it possible to skip this in future */
     uchar enabled;        /* Do we have tributes turned on? */
     uchar bookstock; /* Have we stocked the book? */
     uchar Deathnotice;    /* Did Death notice the book? */
