+++ include/engrave.h
@@ -12,7 +12,7 @@ struct engr {
     struct engr *nxt_engr;
     char *engr_txt;
     xchar engr_x, engr_y;
-    size_t engr_lth; /* for save & restore; not length of text */
+    uint64_t engr_lth; /* for save & restore; not length of text */
     int64_t engr_time;    /* moment engraving was (will be) finished */
     xchar engr_type;
 #define DUST 1
